msm: vidc: Cache bus votes properly when client votes
On targets with ocmem, for each client vote, we cache two copies of
the vote in device->bus_load.  As a result, when triggering an internal
vote (e.g. coming out of power collapse), we aggregate the entries in
device->bus_load and end up voting for double the required bandwidth.

The hfi layer just needs to mirror the client's previous vote.  So just
copy and cache the votes without modification.

Bug: 17502454
Change-Id: Iacca9fb846a05a41630392d9a0c38b40e44a2ad9
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: Ajay Dudani <adudani@codeaurora.org>