qcom: ssr: Free interrupts during subsys_unregister
When a subsystem driver calls subsys_unregister on a subsystem,
the SSR framework removes all instances of the subsystem.
Free the interrupts registered during subsys_register as well
during this time so that the call to request for these IRQs the
next time the subsystem is registered does not fail.

Change-Id: I8163415459c62f4f96ed7c89de9645f9e002c280
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>